requirements/base.txt
Django==1.9.7
djangorestframework==3.4.6
djangorestframework-jwt==1.9.0
django-cors-headers==1.2.0
